requirements.txt
numpy<=1.14.5
pip>=9.0.0
pandas
scikit-learn
scipy
tensorflow>=1.0.0
google-api-python-client
google-cloud-vision
pillow
keras
matplotlib
flask